+++ CHANGELOG.md
@@ -12,3 +12,5 @@
 * renamed cc_custom to cc_ha
 * renamed cc_gwlb_custom to cc_gwlb
 * added bastion source ssh security group option
+* moved all network infrastructure resources (vpc, IGW, NAT Gateway, subnets, route tables, etc.) to unique module terraform-zscc-network-aws
+* added custom subnet definition capabilities via variables cc_subnets, route53_subnets, public_subnets, and workloads_subnets should customer try to override vpc_cidr and the auto cidrsubnet selection becomes incompatible